# machine-history/src/main/resources/META-INF/openapi.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "Machine History Final v2",
    "version" : "1.0.0"
  },
  "paths" : {
    "/machines" : {
      "get" : {
        "summary" : "Gets a list of all machines",
        "operationId" : "getAllMachines",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "The list of machine ID's",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Machine"
              }
            }
          }
        }
      }
    },
    "/machines/{id}" : {
      "get" : {
        "operationId" : "getMachine",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "schema" : {
              "$ref" : "#/definitions/MachineState"
            }
          }
        }
      },
      "parameters" : [ {
        "name" : "id",
        "in" : "path",
        "required" : true,
        "type" : "integer"
      } ]
    },
    "/machines/history/{id}" : {
      "get" : {
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "schema" : {
              "$ref" : "#/definitions/MachineHistory"
            }
          }
        }
      },
      "parameters" : [ {
        "name" : "id",
        "in" : "path",
        "required" : true,
        "type" : "integer"
      } ]
    }
  },
  "definitions" : {
    "MachineState" : {
      "description" : "The representation of the current state of a machine",
      "required" : [ "id", "name", "health" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "description" : "The machine ID",
          "type" : "integer"
        },
        "name" : {
          "description" : "The machine name",
          "type" : "string"
        },
        "health" : {
          "description" : "The machine health",
          "type" : "integer"
        }
      }
    },
    "Machine" : {
      "description" : "A Machine",
      "required" : [ "id", "name" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "format" : "int32",
          "description" : "The Machine ID",
          "type" : "integer"
        },
        "name" : {
          "description" : "The machine name",
          "type" : "string"
        }
      }
    },
    "MaintenanceData" : {
      "description" : "A record of maintenance performed on a machine",
      "required" : [ "id", "date", "health", "mechanic" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "date" : {
          "type" : "string"
        },
        "health" : {
          "type" : "integer"
        },
        "mechanic" : {
          "type" : "string"
        }
      }
    },
    "MachineHistory" : {
      "description" : "The history of a machine",
      "required" : [ "maintenance", "damage" ],
      "type" : "object",
      "properties" : {
        "maintenance" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/MaintenanceData"
          }
        },
        "damage" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DamageData"
          }
        }
      }
    },
    "DamageData" : {
      "required" : [ "id", "amount", "date" ],
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "amount" : {
          "type" : "integer"
        },
        "date" : {
          "type" : "string"
        }
      }
    }
  }
}